feat(#7): implement bytes atoms

+++ eo2js/test/it/runtime-tests.test.js
@@ -15,7 +15,6 @@ const assert = require('assert');
 const exclude = [
   'as-phi-tests',
   'bool-tests',
-  'bytes-tests',
   'cage-tests',
   'cti-test',
   'dataized-tests',
@@ -55,6 +54,14 @@ const allFilesFrom = function(dir) {
 }
 
 /**
+ * This variable allows to skip downloading sources and compilation by eo-maven-plugin if it was
+ * done before. It's useful when you need to run runtime tests multiple times.
+ * Don't forget to set it to TRUE before push!
+ * @type {boolean}
+ */
+const COMPILE = true
+
+/**
  * This test downloads EO tests from objectionary/home repository, parses and assembles them using
  * eo-maven-plugin, transpiles and executes using eo2js.
  */
@@ -64,32 +71,40 @@ describe('runtime tests', function() {
   const project = path.resolve(target, 'project')
   const runtime = path.resolve('../eo2js-runtime')
   before('prepare environment', function() {
-    fs.rmSync(home, {recursive: true, force: true})
-    fs.mkdirSync(project, {recursive: true})
+    if (COMPILE) {
+      fs.rmSync(home, {recursive: true, force: true})
+      fs.mkdirSync(project, {recursive: true})
+    }
   })
   it('should execute all eo-runtime tests', function(done) {
     this.timeout(0)
-    execSync([
-      'git init',
-      'git remote add origin https://github.com/objectionary/home.git',
-      'git config core.sparseCheckout true',
-      'echo tests/org/eolang > .git/info/sparse-checkout',
-      'git pull origin master'
-    ].join(' && '), {cwd: home})
-    console.debug(`Downloaded:\n${
-      allFilesFrom(path.resolve(home, 'tests', 'org', 'eolang'))
-        .map((pth) => path.relative(home, pth))
-        .join(', ')
-    }`)
-    console.debug(`\nExcluded:\n${exclude.join(', ')}`)
-    mvnw(
-      ['register', 'assemble', 'verify'],
-      {home, sources: 'tests', target: 'target'}
-    )
+    if (COMPILE) {
+      execSync([
+        'git init',
+        'git remote add origin https://github.com/objectionary/home.git',
+        'git config core.sparseCheckout true',
+        'echo tests/org/eolang > .git/info/sparse-checkout',
+        'git pull origin master'
+      ].join(' && '), {cwd: home})
+      console.debug(`Downloaded:\n${
+        allFilesFrom(path.resolve(home, 'tests', 'org', 'eolang'))
+          .map((pth) => path.relative(home, pth))
+          .join(', ')
+      }`)
+      console.debug(`\nExcluded:\n${exclude.join(', ')}`)
+      mvnw(
+        ['register', 'assemble', 'verify'],
+        {home, sources: 'tests', target: 'target'}
+      )
+    }
+    if (!COMPILE) {
+      runSync(['link -t', target, '-p project --tests --alone -d', runtime])
+    }
     const log = runSync([
       'test',
       '-t', target,
       '-p project -d', runtime,
+      !COMPILE ? '--alone' : '',
       '--exclude', exclude.join(',')
     ])
     console.debug(log)
